NonstationaryProcessesBase.ParameterProfiles.constantParameter — Function

A function for creating a constant parameter profile, such that if f = constantParamter(x) then f(x)(t) = x

NonstationaryProcessesBase.ParameterProfiles.rampInterval — Method

Define a line over an interval. Saturate before and after this interval

NonstationaryProcessesBase.ParameterProfiles.rampOff — Method

Define a line over an interval. After this interval, saturate, but before, extrapolate.

NonstationaryProcessesBase.ParameterProfiles.rampOn — Method

Define a line over an interval. Before this interval, saturate, but after, extrapolate.

NonstationaryProcessesBase.ParameterProfiles.stepNoise — Function

stepNoise()

Compose a discontinuous function of many square bumps and troughs, following a white noise process. 'stepHeight': the standard deviation of the step height 'T': a tuple, (t0, tmax)

Examples

julia> D = stepNoise();
D([-1, 0, 1])

NonstationaryProcessesBase.ParameterProfiles.unitStep — Function

unitStep(threshold::Real=0.0, baseline::Real=0.0, stepHeight::Real=1.0, stepOpt::Real=1.0)

Construct a constant function of 'x' that undergoes a step of size 'stepHeight' from a 'baseline' at 'threshold'. 'stepOpt' sets the value at 0. A Discontinuous type is returned, which can be called like a normal function but also contains the x-value for which there is a discontinuity.

Examples

julia> D = unitStep();
D([-1, 0, 1])